+++ src/audiounit.rs
@@ -184,7 +184,6 @@ pub trait AudioUnit: Send + Sync + DynClone {
     /// ### Example
     /// ```
     /// use fundsp::hacker::*;
-    /// use num_complex::Complex64;
     /// assert_eq!(pass().response(0, 440.0), Some(Complex64::new(1.0, 0.0)));
     /// ```
     fn response(&mut self, output: usize, frequency: f64) -> Option<Complex64> {
@@ -223,9 +222,9 @@ pub trait AudioUnit: Send + Sync + DynClone {
     /// ```
     /// use fundsp::hacker::*;
     /// assert_eq!(pass().latency(), Some(0.0));
-    /// assert_eq!(tick().latency(), Some(1.0));
+    /// assert_eq!(tick().latency(), Some(0.0));
     /// assert_eq!(sink().latency(), None);
-    /// assert_eq!(lowpass_hz(440.0, 1.0).latency(), Some(0.0));
+    /// assert_eq!(limiter(0.01, 0.01).latency(), Some(441.0));
     /// ```
     fn latency(&mut self) -> Option<f64> {
         if self.outputs() == 0 {